Split out a MaterialBase class ref #12066
